DB changes. Change Roles Id type.

+++ UserDB/dbo/Tables/Users/UserRoles.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[UserRoles]
 (
 	[UserId] BIGINT NOT NULL , 
-    [RoleId] BIGINT NOT NULL, 
+    [RoleId] SMALLINT NOT NULL, 
     [CreatedAt] DATETIME NOT NULL DEFAULT SYSUTCDATETIME(), 
     CONSTRAINT PK_UserRoles PRIMARY KEY ([UserId],[RoleId]),
     CONSTRAINT FK_UserRoles_Users  FOREIGN KEY ([UserId]) REFERENCES [dbo].[Users] (Id),
